requirements.txt
amqp==5.1.1
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==2.4.0
django-queryset-csv==1.1.0
django-rest-multiple-models==2.1.3
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==4.7.2
enum-compat==0.0.3
et-xmlfile==1.1.0
idna==3.2
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.0.1
kombu==5.2.4
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.6
openapi-codec==1.3.2
openpyxl==3.0.10
pandas==1.3.5
prompt-toolkit==3.0.30
psycopg2==2.9.1
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==4.1.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.8.0

